requirements.txt
django==4.1.13
djongo==1.3.6
pymongo==3.12.3
dnspython==2.6.1
python-dotenv
whitenoise
djangorestframework
gunicorn
django-cors-headers
django-extensions